+++ src/lib.rs
@@ -1,305 +1,12 @@
 mod util;
+mod state;
+mod vertex;
 
-use winit::event::{Event, KeyboardInput, VirtualKeyCode, WindowEvent};
-use winit::event_loop::{ControlFlow, EventLoop};
-use winit::window::{Window, WindowBuilder};
-use wgpu::util::DeviceExt;
 use clap::Parser;
-use winit::dpi::PhysicalSize;
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-pub struct Vertex {
-    position: [f32; 3],
-    color: [f32; 3],
-}
-
-impl Vertex {
-    fn desc() -> wgpu::VertexBufferLayout<'static> {
-        wgpu::VertexBufferLayout {
-            array_stride: std::mem::size_of::<Vertex>() as wgpu::BufferAddress,
-            step_mode: wgpu::VertexStepMode::Vertex,
-            attributes: &[
-                wgpu::VertexAttribute {
-                    offset: 0,
-                    shader_location: 0,
-                    format: wgpu::VertexFormat::Float32x3,
-                },
-                wgpu::VertexAttribute {
-                    offset: std::mem::size_of::<[f32; 3]>() as wgpu::BufferAddress,
-                    shader_location: 1,
-                    format: wgpu::VertexFormat::Float32x3,
-                }
-            ],
-        }
-    }
-}
-
-// todo use only for gpu
-const VERTICES: &[Vertex] = &[
-    Vertex { position: [-1.0, -1.0, 0.0], color: [0.0, 0.0, 0.0] },
-    Vertex { position: [1.0, 1.0, 0.0], color: [0.0, 0.0, 0.0] },
-    Vertex { position: [-1.0, 1.0, 0.0], color: [0.0, 0.0, 0.0] },
-    Vertex { position: [1.0, -1.0, 0.0], color: [0.0, 0.0, 0.0] },
-];
-
-const INDICES: &[u16] = &[
-    0, 1, 2,
-    0, 3, 1,
-];
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-struct JuliaUniforms {
-    c: [f32; 2],
-}
-
-struct State {
-    args: Args,
-    surface: wgpu::Surface,
-    device: wgpu::Device,
-    queue: wgpu::Queue,
-    config: wgpu::SurfaceConfiguration,
-    size: PhysicalSize<u32>,
-    window: Window,
-    julia_bind_group: wgpu::BindGroup,
-    render_pipeline: wgpu::RenderPipeline,
-    vertices: Vec<Vertex>, // todo gpu
-    vertex_buffer: wgpu::Buffer,
-    index_buffer: wgpu::Buffer,
-    _num_vertices: u32,
-    num_indices: u32,
-}
-
-impl State {
-    async fn new(window: Window, args: Args) -> Self {
-        // the instance is a handle to our GPU
-        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
-            backends: wgpu::Backends::VULKAN,
-            dx12_shader_compiler: Default::default(),
-        });
-
-        // the surface needs to live as long as the window that created it,
-        // state owns the window so this should be safe
-        let surface = unsafe { instance.create_surface(&window) }.unwrap();
-
-        let adapter = instance.request_adapter(&wgpu::RequestAdapterOptions {
-            power_preference: wgpu::PowerPreference::default(),
-            compatible_surface: Some(&surface),
-            force_fallback_adapter: false,
-        }).await.unwrap();
-
-        let (device, queue) = adapter.request_device(&wgpu::DeviceDescriptor {
-            features: wgpu::Features::empty(),
-            label: None,
-            limits: Default::default(),
-        }, None).await.unwrap();
-
-        // surface
-        let size = window.inner_size();
-        let surface_capabilities = surface.get_capabilities(&adapter);
-        let surface_format = surface_capabilities.formats.iter()
-            .copied()
-            .find(|f| f.is_srgb())
-            .unwrap_or(surface_capabilities.formats[0]);
-        let config = wgpu::SurfaceConfiguration {
-            usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
-            format: surface_format,
-            width: size.width,
-            height: size.height,
-            present_mode: surface_capabilities.present_modes[0],
-            alpha_mode: surface_capabilities.alpha_modes[0],
-            view_formats: vec![],
-        };
-        surface.configure(&device, &config);
-
-        let shader = device.create_shader_module(wgpu::include_wgsl!("shader.wgsl"));
-
-        // uniforms
-        let julia_uniforms = JuliaUniforms { c: args.constant };
-        let julia_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Julia uniform buffer"),
-            contents: bytemuck::cast_slice(&[julia_uniforms]),
-            usage: wgpu::BufferUsages::UNIFORM,
-        });
-        let julia_bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("Julia bind group layout"),
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                }
-            ],
-        });
-        let julia_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("Julia bind group"),
-            layout: &julia_bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: julia_buffer.as_entire_binding(),
-                }
-            ],
-        });
-
-        // render pipeline
-        let render_pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: Some("Render Pipeline Layout"),
-            bind_group_layouts: &[
-                &julia_bind_group_layout,
-            ],
-            push_constant_ranges: &[],
-        });
-        let vertex_state = wgpu::VertexState {
-            module: &shader,
-            entry_point: "vs_main",
-            buffers: &[Vertex::desc()],
-        };
-        let fragment_state_target = [Some(wgpu::ColorTargetState {
-            format: config.format,
-            blend: Some(wgpu::BlendState::REPLACE),
-            write_mask: wgpu::ColorWrites::ALL,
-        })];
-        let fragment_state = Some(wgpu::FragmentState {
-            module: &shader,
-            entry_point: "fs_main",
-            targets: &fragment_state_target,
-        });
-        let primitive_state = wgpu::PrimitiveState {
-            topology: wgpu::PrimitiveTopology::PointList, // todo cpu - point, gpu - triangle
-            strip_index_format: None,
-            front_face: wgpu::FrontFace::Ccw,
-            cull_mode: Some(wgpu::Face::Back),
-            polygon_mode: wgpu::PolygonMode::Fill,
-            unclipped_depth: false,
-            conservative: false,
-        };
-        let multisample_state = wgpu::MultisampleState {
-            count: 1,
-            mask: !0,
-            alpha_to_coverage_enabled: false,
-        };
-        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("Render Pipeline"),
-            layout: Some(&render_pipeline_layout),
-            vertex: vertex_state,
-            fragment: fragment_state,
-            primitive: primitive_state,
-            depth_stencil: None,
-            multisample: multisample_state,
-            multiview: None,
-        });
-
-        // vertex and index buffers
-        // todo cases for cpu and gpu iter
-        let vertices = util::init_vertices();
-
-        let _num_vertices = vertices.len() as u32;
-        let vertex_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Vertex Buffer"),
-            contents: bytemuck::cast_slice(&vertices[..]),
-            usage: wgpu::BufferUsages::VERTEX | wgpu::BufferUsages::COPY_DST,
-        });
-
-        let num_indices = INDICES.len() as u32;
-        let index_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Index Buffer"),
-            contents: bytemuck::cast_slice(INDICES),
-            usage: wgpu::BufferUsages::INDEX,
-        });
-
-        Self {
-            args,
-            window,
-            surface,
-            device,
-            queue,
-            config,
-            size,
-            julia_bind_group,
-            render_pipeline,
-            vertices,
-            vertex_buffer,
-            index_buffer,
-            _num_vertices,
-            num_indices,
-        }
-    }
-
-    pub fn resize(&mut self, new_size: PhysicalSize<u32>) {
-        if new_size.width > 0 && new_size.height > 0 {
-            self.size = new_size;
-            self.config.width = new_size.width;
-            self.config.height = new_size.height;
-            self.surface.configure(&self.device, &self.config);
-        }
-    }
-
-    fn input(&mut self, _event: &WindowEvent) -> bool {
-        false
-    }
-
-    fn update(&mut self) {
-        // todo if cpu iter
-        self.vertices.iter_mut().for_each(|v| {
-           let it = util::julia_iter([v.position[0], v.position[1]], self.args.constant);
-           v.color[2] = it as f32 / util::MAXIMUM_ITERATIONS as f32;
-        });
-
-        self.queue.write_buffer(&self.vertex_buffer, 0, bytemuck::cast_slice(&self.vertices[..]));
-    }
-
-    fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
-        let output = self.surface.get_current_texture()?;
-        let view = output.texture.create_view(&wgpu::TextureViewDescriptor::default());
-        let mut encoder = self.device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-            label: Some("Render Encoder")
-        });
-
-        let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-            label: Some("Render Pass"),
-            color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                view: &view,
-                resolve_target: None,
-                ops: wgpu::Operations {
-                    load: wgpu::LoadOp::Clear(wgpu::Color {
-                        r: 0.0,
-                        g: 0.0,
-                        b: 0.0,
-                        a: 1.0,
-                    }),
-                    store: true,
-                },
-            })],
-            depth_stencil_attachment: None,
-        });
-
-        render_pass.set_pipeline(&self.render_pipeline);
-        render_pass.set_bind_group(0, &self.julia_bind_group, &[]);
-        render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
-        // todo gpu/cpu iter
-        //render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
-        //render_pass.draw_indexed(0..self.num_indices, 0, 0..1);
-        render_pass.draw(0..640_000, 0..1); // 800 * 800 = 640_000
-
-        drop(render_pass);
-
-        self.queue.submit(std::iter::once(encoder.finish()));
-        output.present();
-
-        Ok(())
-    }
-}
 
 #[derive(Parser)]
 #[command(author, version, about, long_about = None)]
-struct Args {
+pub struct Args {
     /// Frames per second limit
     #[arg(long, default_value_t = 60f32)]
     fps: f32,
@@ -309,8 +16,8 @@ struct Args {
     constant: [f32; 2],
 
     /// Perform Julia iteration in the shader
-    #[arg(long, default_value_t = true)]
-    shader_iter: bool,
+    #[arg(long, default_value_t = false)]
+    use_gpu: bool,
 }
 
 impl Args {
@@ -326,54 +33,16 @@ impl Args {
     }
 }
 
-fn match_event(mut state: &mut State, event: Event<()>, control_flow: &mut ControlFlow) {
-    match event {
-        Event::WindowEvent { ref event, window_id }  if window_id == state.window.id() => if !state.input(event) {
-            match_window_event(&mut state, control_flow, event);
-        }
-        Event::RedrawRequested(window_id) if window_id == state.window.id() => {
-            state.update();
-            match state.render() {
-                Ok(_) => {}
-                Err(wgpu::SurfaceError::Lost) => state.resize(state.size),
-                Err(wgpu::SurfaceError::OutOfMemory) => *control_flow = ControlFlow::Exit,
-                Err(e) => eprintln!("{:?}", e),
-            }
-        }
-        Event::MainEventsCleared => state.window.request_redraw(),
-        _ => ()
-    }
-}
-
-fn match_window_event(state: &mut State, control_flow: &mut ControlFlow, event: &WindowEvent) {
-    match event {
-        WindowEvent::CloseRequested => *control_flow = ControlFlow::Exit,
-        WindowEvent::KeyboardInput { input, .. } => match_keyboard_input(control_flow, input),
-        WindowEvent::Resized(physical_size) => state.resize(*physical_size),
-        WindowEvent::ScaleFactorChanged { new_inner_size, .. } => state.resize(**new_inner_size),
-        _ => ()
-    }
-}
-
-fn match_keyboard_input(control_flow: &mut ControlFlow, input: &KeyboardInput) {
-    if let Some(key) = input.virtual_keycode {
-        match key {
-            VirtualKeyCode::Escape => *control_flow = ControlFlow::Exit,
-            _ => ()
-        }
-    }
-}
-
 pub async fn run() {
-    let event_loop = EventLoop::new();
-    let window = WindowBuilder::new()
+    let event_loop = winit::event_loop::EventLoop::new();
+    let window = winit::window::WindowBuilder::new()
         .with_title("Julia")
-        .with_inner_size(PhysicalSize::new(800, 800))
+        .with_inner_size(winit::dpi::PhysicalSize::new(800, 800))
         .build(&event_loop)
         .unwrap();
 
     let args = Args::parse();
-    let mut state = State::new(window, args).await;
+    let mut state = state::State::new(args, window).await;
 
-    event_loop.run(move |event, _, control_flow| match_event(&mut state, event, control_flow));
+    event_loop.run(move |event, _, control_flow| util::match_event(&mut state, event, control_flow));
 }
