+++ src/state.rs
@@ -1,0 +1,264 @@
+use wgpu::util::DeviceExt;
+use crate::util::Vertex;
+use crate::{Args, util};
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+struct JuliaUniforms {
+    c: [f32; 2],
+}
+
+pub struct State {
+    args: Args,
+    surface: wgpu::Surface,
+    device: wgpu::Device,
+    queue: wgpu::Queue,
+    config: wgpu::SurfaceConfiguration,
+    size: winit::dpi::PhysicalSize<u32>,
+    window: winit::window::Window,
+    julia_bind_group: wgpu::BindGroup,
+    render_pipeline: wgpu::RenderPipeline,
+    vertices: Vec<Vertex>,
+    vertex_buffer: wgpu::Buffer,
+    index_buffer: wgpu::Buffer,
+    _num_vertices: u32,
+    num_indices: u32,
+}
+
+impl State {
+    pub async fn new(args: Args, window: winit::window::Window) -> Self {
+        // the instance is a handle to our GPU
+        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
+            backends: wgpu::Backends::VULKAN,
+            dx12_shader_compiler: Default::default(),
+        });
+
+        // the surface needs to live as long as the window that created it,
+        // state owns the window so this should be safe
+        let surface = unsafe { instance.create_surface(&window) }.unwrap();
+
+        let adapter = instance.request_adapter(&wgpu::RequestAdapterOptions {
+            power_preference: wgpu::PowerPreference::default(),
+            compatible_surface: Some(&surface),
+            force_fallback_adapter: false,
+        }).await.unwrap();
+
+        let (device, queue) = adapter.request_device(&wgpu::DeviceDescriptor {
+            features: wgpu::Features::empty(),
+            label: None,
+            limits: Default::default(),
+        }, None).await.unwrap();
+
+        // surface
+        let size = window.inner_size();
+        let surface_capabilities = surface.get_capabilities(&adapter);
+        let surface_format = surface_capabilities.formats.iter()
+            .copied()
+            .find(|f| f.is_srgb())
+            .unwrap_or(surface_capabilities.formats[0]);
+        let config = wgpu::SurfaceConfiguration {
+            usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
+            format: surface_format,
+            width: size.width,
+            height: size.height,
+            present_mode: surface_capabilities.present_modes[0],
+            alpha_mode: surface_capabilities.alpha_modes[0],
+            view_formats: vec![],
+        };
+        surface.configure(&device, &config);
+
+        let shader = device.create_shader_module(wgpu::include_wgsl!("shader.wgsl"));
+
+        // uniforms
+        let julia_uniforms = JuliaUniforms { c: args.constant };
+        let julia_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("Julia uniform buffer"),
+            contents: bytemuck::cast_slice(&[julia_uniforms]),
+            usage: wgpu::BufferUsages::UNIFORM,
+        });
+        let julia_bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+            label: Some("Julia bind group layout"),
+            entries: &[
+                wgpu::BindGroupLayoutEntry {
+                    binding: 0,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Buffer {
+                        ty: wgpu::BufferBindingType::Uniform,
+                        has_dynamic_offset: false,
+                        min_binding_size: None,
+                    },
+                    count: None,
+                }
+            ],
+        });
+        let julia_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+            label: Some("Julia bind group"),
+            layout: &julia_bind_group_layout,
+            entries: &[
+                wgpu::BindGroupEntry {
+                    binding: 0,
+                    resource: julia_buffer.as_entire_binding(),
+                }
+            ],
+        });
+
+        // render pipeline
+        let render_pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+            label: Some("Render Pipeline Layout"),
+            bind_group_layouts: &[
+                &julia_bind_group_layout,
+            ],
+            push_constant_ranges: &[],
+        });
+        let vertex_state = wgpu::VertexState {
+            module: &shader,
+            entry_point: "vs_main",
+            buffers: &[Vertex::desc()],
+        };
+        let fragment_state_target = [Some(wgpu::ColorTargetState {
+            format: config.format,
+            blend: Some(wgpu::BlendState::REPLACE),
+            write_mask: wgpu::ColorWrites::ALL,
+        })];
+        let fragment_state = Some(wgpu::FragmentState {
+            module: &shader,
+            entry_point: "fs_main",
+            targets: &fragment_state_target,
+        });
+        let primitive_state = wgpu::PrimitiveState {
+            topology: if args.use_gpu { wgpu::PrimitiveTopology::TriangleList } else { wgpu::PrimitiveTopology::PointList },
+            strip_index_format: None,
+            front_face: wgpu::FrontFace::Ccw,
+            cull_mode: Some(wgpu::Face::Back),
+            polygon_mode: wgpu::PolygonMode::Fill,
+            unclipped_depth: false,
+            conservative: false,
+        };
+        let multisample_state = wgpu::MultisampleState {
+            count: 1,
+            mask: !0,
+            alpha_to_coverage_enabled: false,
+        };
+        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
+            label: Some("Render Pipeline"),
+            layout: Some(&render_pipeline_layout),
+            vertex: vertex_state,
+            fragment: fragment_state,
+            primitive: primitive_state,
+            depth_stencil: None,
+            multisample: multisample_state,
+            multiview: None,
+        });
+
+        // vertex and index buffers
+        let vertices = Vertex::init_vertices(args.use_gpu);
+        let _num_vertices = vertices.len() as u32;
+        let vertex_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("Vertex Buffer"),
+            contents: bytemuck::cast_slice(&vertices[..]),
+            usage: wgpu::BufferUsages::VERTEX | wgpu::BufferUsages::COPY_DST,
+        });
+
+        let indices = Vertex::init_indices();
+        let num_indices = indices.len() as u32;
+        let index_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("Index Buffer"),
+            contents: bytemuck::cast_slice(indices.as_slice()),
+            usage: wgpu::BufferUsages::INDEX,
+        });
+
+        Self {
+            args,
+            window,
+            surface,
+            device,
+            queue,
+            config,
+            size,
+            julia_bind_group,
+            render_pipeline,
+            vertices,
+            vertex_buffer,
+            index_buffer,
+            _num_vertices,
+            num_indices,
+        }
+    }
+
+    pub fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
+        if new_size.width > 0 && new_size.height > 0 {
+            self.size = new_size;
+            self.config.width = new_size.width;
+            self.config.height = new_size.height;
+            self.surface.configure(&self.device, &self.config);
+        }
+    }
+
+    pub fn input(&mut self, _event: &winit::event::WindowEvent) -> bool {
+        false
+    }
+
+    pub fn update(&mut self) {
+        if self.args.use_gpu { return; }
+
+        self.vertices.iter_mut().for_each(|v| {
+            let it = util::julia_iter(v.position(), self.args.constant) as f32;
+            let b = it / util::MAXIMUM_ITERATIONS as f32;
+            v.set_color([0f32, 0f32, b]);
+        });
+
+        self.queue.write_buffer(&self.vertex_buffer, 0, bytemuck::cast_slice(&self.vertices[..]));
+    }
+
+    pub fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
+        let output = self.surface.get_current_texture()?;
+        let view = output.texture.create_view(&wgpu::TextureViewDescriptor::default());
+        let mut encoder = self.device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
+            label: Some("Render Encoder")
+        });
+
+        let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
+            label: Some("Render Pass"),
+            color_attachments: &[Some(wgpu::RenderPassColorAttachment {
+                view: &view,
+                resolve_target: None,
+                ops: wgpu::Operations {
+                    load: wgpu::LoadOp::Clear(wgpu::Color {
+                        r: 0.0,
+                        g: 0.0,
+                        b: 0.0,
+                        a: 1.0,
+                    }),
+                    store: true,
+                },
+            })],
+            depth_stencil_attachment: None,
+        });
+
+        render_pass.set_pipeline(&self.render_pipeline);
+        render_pass.set_bind_group(0, &self.julia_bind_group, &[]);
+        render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
+
+        if self.args.use_gpu {
+            render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
+            render_pass.draw_indexed(0..self.num_indices, 0, 0..1);
+        } else {
+            render_pass.draw(0..640_000, 0..1);
+        }
+
+        drop(render_pass);
+
+        self.queue.submit(std::iter::once(encoder.finish()));
+        output.present();
+
+        Ok(())
+    }
+
+    pub fn window(&self) -> &winit::window::Window {
+        &self.window
+    }
+
+    pub fn size(&self) -> winit::dpi::PhysicalSize<u32> {
+        self.size
+    }
+}
